feat: Make Manager Agent compatible with Execution Agent
Clones the new Execution Agent repository and modifies the Manager Agent to ensure compatibility.

The key changes include:
- Updating the `CreateOrderRequest` model in `app/models.py` to match the Execution Agent's contract.
  - `quantity` is changed from `Decimal` to `int`.
  - `price` is changed from `Optional[Decimal]` to `Optional[float]`.
- Modifying `app/risk_manager.py` to explicitly cast the calculated `position_size` to an integer, ensuring the data passed to the model is of the correct type.

These changes resolve the data type mismatch that would have caused validation errors, allowing the two services to communicate and execute trades correctly. All existing tests pass, confirming that these changes do not introduce regressions.

+++ app/models.py
@@ -188,8 +188,8 @@ class CreateOrderRequest(BaseModel):
     symbol: str
     side: OrderSide
     order_type: OrderType
-    price: Optional[Decimal] = None
-    quantity: Decimal
+    price: Optional[float] = None
+    quantity: int
     time_in_force: TimeInForce = TimeInForce.GTC
